Added pod to AuthDirectory and fixed up most of the failed tests. The only remaining failed tests are in authdirectory-coverage, and are related to the preferred value for an uninitialized attribute. Currently many of them default an empty string, but a preference for undef has been encoded in the tests. This will require some jiggering in the backend profile service to implement properly, and is work for another day.
   The AuthUser->user_id attribute has been modified so that once it
has been set, it cannot be reset with the user_id() getter/setter.
Of course, nothing prevents anyone from accessing the underlying hash
and rewriting it. A proper fix requires that we update the back
end django-piston service so that updates cannot overwrite the
primary key of an existing record. It appears that django-piston
is not distinguishing between POST and PUT, and it will need some
back end coding as well.